referenciaSanluis-Version 1.0.2/requirements.txt
mysql-connector-python~=9.0.0
tk
pandas~=2.2.2
matplotlib~=3.9.1